backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3
gunicorn>=21.2
psycopg2-binary>=2.9
dj-database-url>=2.1
whitenoise>=6.6
